requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scikit-learn
umap-learn
joblib
matplotlib
plotly